#88 feat: BE plugins bundler
